move module category screen
