+++ app/src/main/java/com/jin/honey/feature/orderdetail/ui/OrderDetailScreen.kt
@@ -37,7 +37,7 @@ import com.jin.feature.ui.state.SearchState
 import com.jin.feature.ui.state.UiState
 import com.jin.honey.R
 import com.jin.ui.cart.content.CartOptionModifyBottomSheet
-import com.jin.honey.feature.home.ui.content.headercontent.LocationSearchBottomSheet
+import com.jin.ui.address.LocationSearchBottomSheet
 import com.jin.domain.order.model.Order
 import com.jin.domain.order.model.PayPrice
 import com.jin.domain.order.model.PaymentState
