+++ app/src/main/res/drawable/ic_shopping_basket.xml
@@ -1,0 +1,36 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M69.82,200.73c-4.47,0 -8.93,-1.71 -12.34,-5.11c-6.82,-6.81 -6.82,-17.87 0,-24.68L208.75,19.66c6.82,-6.82 17.86,-6.82 24.68,0c6.82,6.81 6.82,17.87 0,24.68L82.16,195.61C78.75,199.02 74.28,200.73 69.82,200.73z"
+      android:fillColor="#736056"/>
+  <path
+      android:pathData="M442.18,200.73c-4.47,0 -8.93,-1.71 -12.34,-5.11L278.57,44.34c-6.82,-6.81 -6.82,-17.87 0,-24.68c6.82,-6.82 17.86,-6.82 24.68,0l151.27,151.27c6.82,6.81 6.82,17.87 0,24.68C451.11,199.02 446.65,200.73 442.18,200.73z"
+      android:fillColor="#665247"/>
+  <path
+      android:pathData="M25.94,229.82l33.44,223.88c4.23,25.35 25.95,43.76 51.65,43.76h295.06c25.89,0 48.18,-19.32 51.84,-44.96l28.48,-222.68H25.94z"
+      android:fillColor="#fa0050"/>
+  <path
+      android:pathData="M256,497.45h150.09c25.89,0 48.18,-19.32 51.84,-44.96l28.48,-222.68H256V497.45z"
+      android:fillColor="#CE0244"/>
+  <path
+      android:pathData="M17.45,229.82C7.83,229.82 0,221.99 0,212.36v-34.91C0,167.83 7.83,160 17.45,160h477.09c9.62,0 17.45,7.83 17.45,17.45v34.91c0,9.62 -7.83,17.45 -17.45,17.45H17.45z"
+      android:fillColor="#D7EBFF"/>
+  <path
+      android:pathData="M151.27,427.64c-9.64,0 -17.45,-7.81 -17.45,-17.45V305.45c0,-9.64 7.81,-17.45 17.45,-17.45c9.64,0 17.45,7.81 17.45,17.45v104.73C168.73,419.82 160.91,427.64 151.27,427.64z"
+      android:fillColor="#D7EBFF"/>
+  <path
+      android:pathData="M360.73,427.64c-9.64,0 -17.45,-7.81 -17.45,-17.45V305.45c0,-9.64 7.81,-17.45 17.45,-17.45s17.45,7.81 17.45,17.45v104.73C378.18,419.82 370.37,427.64 360.73,427.64z"
+      android:fillColor="#BDDEFF"/>
+  <path
+      android:pathData="M256,427.64c-9.64,0 -17.45,-7.81 -17.45,-17.45V305.45c0,-9.64 7.81,-17.45 17.45,-17.45c9.64,0 17.45,7.81 17.45,17.45v104.73C273.45,419.82 265.64,427.64 256,427.64z"
+      android:fillColor="#D7EBFF"/>
+  <path
+      android:pathData="M273.45,410.18V305.45c0,-9.64 -7.81,-17.45 -17.45,-17.45v139.64C265.64,427.64 273.45,419.82 273.45,410.18z"
+      android:fillColor="#BDDEFF"/>
+  <path
+      android:pathData="M494.55,160H256v69.82h238.54c9.62,0 17.45,-7.83 17.45,-17.45v-34.91C512,167.83 504.17,160 494.55,160z"
+      android:fillColor="#BDDEFF"/>
+</vector>
