+++ ui/src/main/res/drawable/ic_chinese.xml
@@ -1,0 +1,89 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M231.62,270.97c-3.29,-6.63 -10.11,-11.19 -18.01,-11.19c-2.81,0 -5.48,0.58 -7.91,1.62c-3.19,-6.89 -10.15,-11.68 -18.24,-11.68s-15.05,4.79 -18.24,11.68c-2.43,-1.04 -5.1,-1.62 -7.91,-1.62c-7.9,0 -14.72,4.57 -18.01,11.19c-30.83,5.73 -54.18,32.75 -54.18,65.24c0,36.66 29.72,66.37 66.37,66.37h63.92c36.66,0 66.37,-29.72 66.37,-66.37C285.79,303.72 262.45,276.7 231.62,270.97z"
+      android:fillColor="#FFEA92"/>
+  <path
+      android:pathData="M231.62,270.97c-3.29,-6.63 -10.11,-11.19 -18.01,-11.19c-2.81,0 -5.48,0.58 -7.91,1.62c-3.19,-6.89 -10.15,-11.68 -18.24,-11.68s-15.05,4.79 -18.24,11.68c-2.43,-1.04 -5.1,-1.62 -7.91,-1.62c-7.9,0 -14.72,4.57 -18.01,11.19c-30.83,5.73 -54.18,32.75 -54.18,65.24c0,36.66 29.72,66.37 66.37,66.37h63.92c36.66,0 66.37,-29.72 66.37,-66.37C285.79,303.72 262.45,276.7 231.62,270.97z"
+      android:fillColor="#FFEA92"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M252.59,380.95h-63.91c-36.66,0 -66.37,-29.72 -66.37,-66.37c0,-16.06 5.71,-30.77 15.2,-42.24c-27.9,7.85 -48.37,33.47 -48.37,63.88c0,36.66 29.72,66.37 66.37,66.37h63.92c20.61,0 39.03,-9.39 51.2,-24.13C264.88,380.07 258.84,380.95 252.59,380.95z"
+      android:strokeAlpha="0.1"
+      android:fillAlpha="0.1"/>
+  <path
+      android:pathData="M166.21,291.65c-2.04,0 -4.07,-0.78 -5.63,-2.33l-3.62,-3.62c-3.11,-3.11 -3.11,-8.15 0,-11.26c3.11,-3.11 8.15,-3.11 11.26,0l3.62,3.62c3.11,3.11 3.11,8.15 -0,11.26C170.28,290.87 168.24,291.65 166.21,291.65z"
+      android:fillColor="#E5D07A"/>
+  <path
+      android:pathData="M209.78,291.16c-2.04,0 -4.07,-0.78 -5.63,-2.33c-3.11,-3.11 -3.11,-8.15 0,-11.26l3.3,-3.3c3.11,-3.11 8.15,-3.11 11.26,0c3.11,3.11 3.11,8.15 0,11.26l-3.3,3.3C213.85,290.39 211.81,291.16 209.78,291.16z"
+      android:fillColor="#E5D07A"/>
+  <path
+      android:pathData="M188.06,287.8c-4.39,0 -7.96,-3.56 -7.96,-7.96v-6.15c0,-4.4 3.56,-7.96 7.96,-7.96c4.4,0 7.96,3.56 7.96,7.96v6.15C196.02,284.24 192.46,287.8 188.06,287.8z"
+      android:fillColor="#E5D07A"/>
+  <path
+      android:pathData="M366.64,249.74c-3.29,-6.63 -10.11,-11.19 -18.01,-11.19c-2.81,0 -5.48,0.58 -7.91,1.62c-3.19,-6.89 -10.15,-11.68 -18.24,-11.68c-8.09,0 -15.05,4.79 -18.24,11.68c-2.43,-1.04 -5.1,-1.62 -7.91,-1.62c-7.9,0 -14.72,4.57 -18.01,11.19c-30.83,5.73 -54.18,32.75 -54.18,65.24c0,36.66 29.72,66.37 66.37,66.37h63.91c36.66,0 66.37,-29.72 66.37,-66.37C420.82,282.49 397.47,255.47 366.64,249.74z"
+      android:fillColor="#F2B47C"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M387.35,359.73h-63.91c-36.66,0 -66.37,-29.72 -66.37,-66.37c0,-16.01 5.68,-30.69 15.13,-42.15c-27.73,7.96 -48.03,33.49 -48.03,63.78c0,36.66 29.72,66.37 66.37,66.37h63.91c20.65,0 39.1,-9.43 51.27,-24.22C399.88,358.81 393.72,359.73 387.35,359.73z"
+      android:strokeAlpha="0.1"
+      android:fillAlpha="0.1"/>
+  <path
+      android:pathData="M301.23,270.42c-2.04,0 -4.07,-0.78 -5.63,-2.33l-3.62,-3.62c-3.11,-3.11 -3.11,-8.15 0,-11.26c3.11,-3.11 8.15,-3.11 11.26,0l3.62,3.62c3.11,3.11 3.11,8.15 0,11.26C305.3,269.64 303.27,270.42 301.23,270.42z"
+      android:fillColor="#D89B68"/>
+  <path
+      android:pathData="M344.8,269.93c-2.04,0 -4.07,-0.78 -5.63,-2.33c-3.11,-3.11 -3.11,-8.15 0,-11.26l3.3,-3.3c3.11,-3.11 8.15,-3.11 11.26,0c3.11,3.11 3.11,8.15 0,11.26l-3.3,3.3C348.87,269.16 346.84,269.93 344.8,269.93z"
+      android:fillColor="#D89B68"/>
+  <path
+      android:pathData="M323.08,266.57c-4.39,0 -7.96,-3.56 -7.96,-7.96v-6.15c0,-4.4 3.56,-7.96 7.96,-7.96c4.39,0 7.96,3.56 7.96,7.96v6.15C331.04,263.01 327.48,266.57 323.08,266.57z"
+      android:fillColor="#D89B68"/>
+  <path
+      android:pathData="M445.09,352.7c0,6.45 -5.23,11.67 -11.67,11.67H78.58c-6.45,0 -11.67,-5.23 -11.67,-11.67v-21.01c0,-6.45 5.23,-11.67 11.67,-11.67h354.83c6.45,0 11.67,5.23 11.67,11.67V352.7z"
+      android:fillColor="#F7EEDF"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M111.27,352.7v-21.01c0,-6.45 5.23,-11.67 11.67,-11.67H78.58c-6.45,0 -11.67,5.23 -11.67,11.67v21.01c0,6.45 5.23,11.67 11.67,11.67h44.36C116.5,364.37 111.27,359.14 111.27,352.7z"
+      android:strokeAlpha="0.1"
+      android:fillAlpha="0.1"/>
+  <path
+      android:pathData="M226.54,222.73h-26.76c-4.39,0 -7.96,-3.56 -7.96,-7.96s3.56,-7.96 7.96,-7.96h26.76c15.94,0 28.91,-12.97 28.91,-28.92c0,-15.94 -12.97,-28.91 -28.91,-28.91h-27.45c-4.4,0 -7.96,-3.56 -7.96,-7.96c0,-4.4 3.56,-7.96 7.96,-7.96h27.45c24.72,0 44.83,20.11 44.83,44.83C271.37,202.62 251.26,222.73 226.54,222.73z"
+      android:fillColor="#FCE9C3"/>
+  <path
+      android:pathData="M172.31,148.98h-32.21c-27.11,0 -49.17,-22.06 -49.17,-49.17c0,-27.11 22.06,-49.17 49.17,-49.17h36.36c4.4,0 7.96,3.56 7.96,7.96c0,4.39 -3.56,7.96 -7.96,7.96h-36.36c-18.33,0 -33.25,14.92 -33.25,33.25s14.92,33.25 33.25,33.25h32.21c4.39,0 7.96,3.56 7.96,7.96C180.27,145.42 176.71,148.98 172.31,148.98z"
+      android:fillColor="#F2D8A0"/>
+  <path
+      android:pathData="M349.21,0h-23.05c-4.4,0 -7.96,3.56 -7.96,7.96c0,4.4 3.56,7.96 7.96,7.96h23.05c14.03,0 25.45,11.42 25.45,25.45c0,14.03 -11.42,25.45 -25.45,25.45h-66.3c-19.82,0 -35.95,16.13 -35.95,35.95v6.65c0,4.39 3.56,7.96 7.96,7.96c4.4,0 7.96,-3.56 7.96,-7.96v-6.65c0,-11.05 8.99,-20.03 20.03,-20.03h66.3c22.81,0 41.37,-18.56 41.37,-41.37C390.57,18.56 372.02,0 349.21,0z"
+      android:fillColor="#F2D8A0"/>
+  <path
+      android:pathData="M467.85,500.33c0,6.45 -5.23,11.67 -11.67,11.67H55.82c-6.45,0 -11.67,-5.23 -11.67,-11.67V362.6c0,-6.45 5.23,-11.67 11.67,-11.67h400.35c6.45,0 11.67,5.23 11.67,11.67v137.73H467.85z"
+      android:fillColor="#C69C6D"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M88.34,500.33V362.6c0,-6.45 5.23,-11.67 11.67,-11.67H55.82c-6.45,0 -11.67,5.23 -11.67,11.67v137.73c0,6.45 5.23,11.67 11.67,11.67h44.19C93.57,512 88.34,506.77 88.34,500.33z"
+      android:strokeAlpha="0.1"
+      android:fillAlpha="0.1"/>
+  <path
+      android:pathData="M318.08,468.09H44.15v15.92h273.92c4.39,0 7.96,-3.56 7.96,-7.96C326.04,471.65 322.47,468.09 318.08,468.09z"
+      android:fillColor="#A67C52"/>
+  <path
+      android:pathData="M135.97,438.36H44.15v15.92h91.82c4.39,0 7.96,-3.56 7.96,-7.96C143.93,441.93 140.37,438.36 135.97,438.36z"
+      android:fillColor="#A67C52"/>
+  <path
+      android:pathData="M100.49,408.64H44.15v15.92h56.34c4.39,0 7.96,-3.56 7.96,-7.96C108.44,412.21 104.88,408.64 100.49,408.64z"
+      android:fillColor="#A67C52"/>
+  <path
+      android:pathData="M135.97,378.92H44.15v15.92h91.82c4.39,0 7.96,-3.56 7.96,-7.96C143.93,382.48 140.37,378.92 135.97,378.92z"
+      android:fillColor="#A67C52"/>
+  <path
+      android:pathData="M390.85,468.09h-36.7c-4.4,0 -7.96,3.56 -7.96,7.96c0,4.4 3.56,7.96 7.96,7.96h36.7c4.4,0 7.96,-3.56 7.96,-7.96C398.81,471.65 395.25,468.09 390.85,468.09z"
+      android:fillColor="#A67C52"/>
+  <path
+      android:pathData="M467.85,468.09h-53.65c-4.4,0 -7.96,3.56 -7.96,7.96c0,4.4 3.56,7.96 7.96,7.96h53.65V468.09z"
+      android:fillColor="#A67C52"/>
+  <path
+      android:pathData="M158.87,394.84h37.15c4.39,0 7.96,-3.56 7.96,-7.96c0,-4.4 -3.56,-7.96 -7.96,-7.96H158.87c-4.39,0 -7.96,3.56 -7.96,7.96C150.91,391.27 154.48,394.84 158.87,394.84z"
+      android:fillColor="#A67C52"/>
+</vector>
