+++ ui/src/main/res/drawable/ic_crown.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="32"
+    android:viewportHeight="32">
+  <path
+      android:pathData="M16,10.938c-1.375,0 -2.5,-1.125 -2.5,-2.5 0,-1.406 1.125,-2.5 2.5,-2.5s2.5,1.094 2.5,2.5c0,1.375 -1.125,2.5 -2.5,2.5zM6.031,9.906c1.094,0 1.969,0.906 1.969,2 0,1.125 -0.875,2 -1.969,2 -1.125,0 -2.031,-0.875 -2.031,-2 0,-1.094 0.906,-2 2.031,-2zM26.031,9.906c1.094,0 1.969,0.906 1.969,2 0,1.125 -0.875,2 -1.969,2 -1.125,0 -2.031,-0.875 -2.031,-2 0,-1.094 0.906,-2 2.031,-2zM8.219,23.719l-1.656,-9.063c0.5,-0.094 0.969,-0.375 1.344,-0.688 1.031,0.938 2.344,1.844 3.594,1.844 1.5,0 2.719,-2.313 3.563,-4.25 0.281,0.094 0.625,0.188 0.938,0.188s0.656,-0.094 0.938,-0.188c0.844,1.938 2.063,4.25 3.563,4.25 1.25,0 2.563,-0.906 3.594,-1.844 0.375,0.313 0.844,0.594 1.344,0.688l-1.656,9.063h-15.563zM7.875,24.5h16.25v1.531h-16.25v-1.531z"
+      android:fillColor="#000000"/>
+</vector>
