+++ ui/src/main/res/drawable/ic_dessert.xml
@@ -1,0 +1,48 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="128"
+    android:viewportHeight="128">
+  <path
+      android:pathData="M97.9,105.05c0.39,-0.59 11.26,2.12 11.26,2.12s5.84,-0.94 7.81,-3.5s1.55,-6.96 0.96,-8.73c-0.59,-1.77 -2.33,-3.06 -2.33,-3.06s0.79,-13.76 0.2,-26.35c-0.59,-12.58 -1.38,-18.29 -1.38,-18.29s3.79,-2.9 3.39,-8.51s-5.15,-7.12 -7.69,-7.13c-3.57,-0.01 -6.17,2.4 -6.17,2.4s0.48,-7.5 -6.25,-9.53c-4.25,-1.28 -6.57,2.56 -6.57,2.56s0.84,-7.6 -5.37,-9.69c-4.92,-1.65 -7.29,2.8 -7.29,2.8l-18.7,-5.81s-3.77,-3.41 -7.17,-0.6c-2.8,2.32 -2.8,5.69 -2.8,5.69s-9.29,8.73 -18.34,19.94s-18.9,25.77 -20.1,29.31c-0.71,2.1 -0.45,4.93 -0.45,4.93s7.56,10.74 7.72,14.66c0.16,3.92 35.07,11.75 35.23,13.16c0.17,1.42 44.04,3.63 44.04,3.63z"
+      android:fillColor="#ffccc1"/>
+  <path
+      android:pathData="M14.35,112.73s-0.65,3.35 -0.54,6.06c0.11,2.71 0,4 1.73,4.76s34.3,-3.35 53.67,-6.6c19.37,-3.25 41.33,-10.06 41.33,-10.06s1.05,-21.97 0.08,-33.25c-1.08,-12.51 -2.49,-14.71 -6.82,-16.76c-4.33,-2.06 -12.16,0.78 -16.3,2.12c-14,4.51 -38.41,10.87 -65.47,12.29c-8.89,0.47 -11.13,0.36 -11.13,0.36s-0.09,1.03 0.12,4.42c0.28,4.7 0.94,11.95 1.01,12.82c0.1,1.51 2.32,23.84 2.32,23.84z"
+      android:fillColor="#df7c1a"/>
+  <path
+      android:pathData="M58.82,70.32c15.43,-3.13 35.6,-9.63 38.84,-10.06c3.25,-0.43 7.14,-0.22 8.55,5.95c1.41,6.17 1.41,23.8 1.3,26.83s-0.43,9.52 -0.43,9.52s-12.55,5.95 -42.85,10.71c-30.3,4.76 -50.42,5.52 -50.42,5.52L11,76.05s22.72,-0.65 47.82,-5.73z"
+      android:fillColor="#fea32e"/>
+  <path
+      android:pathData="M39.02,88.61c14.23,-2.01 30.95,-5.19 46.74,-9.85c8.8,-2.59 21.18,-7.47 21.18,-7.47s0.2,2.38 0.37,4.97s0.31,7.81 0.31,7.81s-16.77,7.9 -46.31,13.2c-29.54,5.3 -48.49,6.28 -48.49,6.28l-0.79,-11.94c-0.01,0 10.87,-0.73 26.99,-3z"
+      android:fillColor="#ffccc1"/>
+  <path
+      android:pathData="M60.77,95.85c14.72,-2.6 14.49,-14.26 14.49,-14.26s-10.17,2.44 -15.14,3.44c-4.1,0.82 -14.48,2.57 -14.48,2.57s3.75,10.26 15.13,8.25z"
+      android:fillColor="#fe4f20"/>
+  <path
+      android:pathData="M107.62,84.06c0.11,-0.65 -2.01,-9 -13.74,-6.49c-12.27,2.62 -14.01,15.69 -13.39,15.55c0.74,-0.18 7.85,-2.05 14.62,-4.25c7.49,-2.45 12.51,-4.81 12.51,-4.81z"
+      android:fillColor="#fe4f20"/>
+  <path
+      android:pathData="M13.71,103.5s2.38,-11.09 14.97,-11.84c14.5,-0.87 15.93,8.34 15.93,8.34s-11.61,1.78 -16.04,2.26c-7.16,0.79 -14.86,1.24 -14.86,1.24z"
+      android:fillColor="#fe4f20"/>
+  <path
+      android:pathData="M18.55,64.78c-0.54,1.73 2.9,1.18 8.12,0.92c8.87,-0.44 23.95,-2.34 40.5,-6.7c26.29,-6.92 39.14,-11.73 39.14,-12.6c0,-0.87 -19.37,-13.85 -29.54,-16.99s-23.26,-8.44 -23.26,-7.9c0,0.54 -12.61,12.65 -18.72,20.34c-6.26,7.91 -15.5,20.61 -16.24,22.93z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M69.32,46.62c14.53,-0.81 17.85,-7.68 17.64,-17.53c-0.21,-9.43 -8.55,-25.86 -18.18,-25.54c-8.98,0.3 -17.2,14.39 -18.5,24.67s3.46,19.27 19.04,18.4z"
+      android:fillColor="#ff5117"/>
+  <path
+      android:pathData="M67.72,27.33c-0.61,0.11 -2.24,3.37 -2.13,5.29c0.12,2.01 1.09,3.16 2.92,2.98c1.71,-0.17 2.37,-1.52 2.13,-3.47s-2.25,-4.92 -2.92,-4.8z"
+      android:fillColor="#feffff"/>
+  <path
+      android:pathData="M78.42,30.91c-0.97,-0.24 -3.28,3.4 -3.4,5.47c-0.09,1.58 0.97,2.68 2.55,2.68c1.58,0 2.43,-1.22 2.55,-2.8s-0.85,-5.14 -1.7,-5.35z"
+      android:fillColor="#feffff"/>
+  <path
+      android:pathData="M57.14,30.43c-0.91,0.18 -1.81,4.13 -1.4,5.9c0.24,1.03 1.16,2.31 2.86,1.82s1.95,-2.86 1.4,-4.07c-0.55,-1.22 -1.78,-3.87 -2.86,-3.65z"
+      android:fillColor="#feffff"/>
+  <path
+      android:pathData="M63.53,14.68c-0.82,-0.21 -3.28,2.19 -3.83,4.5c-0.48,2.03 0.36,3.22 1.95,3.59c1.58,0.36 2.92,-0.61 3.16,-2.37c0.22,-1.66 -0.31,-5.48 -1.28,-5.72z"
+      android:fillColor="#feffff"/>
+  <path
+      android:pathData="M72.65,16.69c-0.73,0.3 -1.28,4.44 -0.55,6.2c0.78,1.89 2.54,1.94 3.59,1.46c1.46,-0.67 1.62,-2.02 1.22,-3.34c-0.5,-1.59 -3.62,-4.59 -4.26,-4.32z"
+      android:fillColor="#feffff"/>
+</vector>
