+++ ui/src/main/res/drawable/ic_environment.xml
@@ -1,0 +1,51 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M445.43,244.56c0,120.35 -97.03,217.9 -216.73,217.9S11.98,364.9 11.98,244.56S109.01,26.65 228.71,26.65S445.43,124.21 445.43,244.56z"
+      android:fillColor="#8AC6B6"/>
+  <path
+      android:pathData="M309.3,244.56c0,120.35 -36.09,217.9 -80.6,217.9s-80.6,-97.56 -80.6,-217.9s36.09,-217.9 80.6,-217.9S309.3,124.21 309.3,244.56z"
+      android:fillColor="#9FD7C5"/>
+  <path
+      android:pathData="M225.68,152.3c-61.32,0 -119.7,-12.56 -172.73,-35.22c-25.77,35.84 -40.97,79.87 -40.97,127.48c0,120.35 97.03,217.9 216.73,217.9s216.73,-97.56 216.73,-217.9c0,-48.5 -15.77,-93.29 -42.42,-129.49C348.76,138.99 288.78,152.3 225.68,152.3z"
+      android:fillColor="#86BAA8"/>
+  <path
+      android:pathData="M17.15,292.01c21.55,97.52 108.08,170.45 211.56,170.45c40.08,0 77.62,-10.95 109.82,-30.03c-6.4,-7.93 -11.18,-13.75 -12.5,-15.22c59.09,-44.51 97.36,-115.46 97.47,-195.47c-97.44,-0.21 -180.61,55.24 -219.51,138.4C198.61,359.93 86.66,371.51 17.15,292.01z"
+      android:fillColor="#74A58E"/>
+  <path
+      android:pathData="M309.3,244.56c0,-35.5 -3.14,-69.01 -8.71,-98.62c-24.35,4.18 -49.38,6.37 -74.92,6.37c-23.49,0 -46.54,-1.86 -69.04,-5.41c-5.46,29.38 -8.53,62.54 -8.53,97.66c0,120.35 36.09,217.9 80.6,217.9S309.3,364.9 309.3,244.56z"
+      android:fillColor="#BAEEC3"/>
+  <path
+      android:pathData="M225.68,266.66c-73.97,0 -145.59,-10.03 -213.59,-28.79c-0.07,2.22 -0.11,4.45 -0.11,6.69c0,120.35 97.03,217.9 216.73,217.9s216.73,-97.56 216.73,-217.9c0,-2.8 -0.06,-5.59 -0.16,-8.36C375.49,256.04 301.82,266.66 225.68,266.66z"
+      android:fillColor="#63A097"/>
+  <path
+      android:pathData="M228.71,462.46c43.9,0 79.6,-94.9 80.58,-212.96c-45.93,24.18 -82.93,62.83 -105.3,110.64c-2.03,-0.08 -19.32,1.52 -43.78,-0.75C174.42,421.26 199.78,462.46 228.71,462.46z"
+      android:fillColor="#9FDDAA"/>
+  <path
+      android:pathData="M228.71,462.46c40.08,0 77.62,-10.95 109.82,-30.03c-6.4,-7.93 -11.18,-13.75 -12.5,-15.22c53.99,-40.66 90.58,-103.4 96.58,-174.96c-44.65,11.28 -90.79,18.82 -138.07,22.26c-34.49,23.91 -62.36,56.74 -80.56,95.63c-5.37,-0.21 -117.32,11.36 -186.83,-68.13C38.69,389.54 125.22,462.46 228.71,462.46z"
+      android:fillColor="#568C82"/>
+  <path
+      android:pathData="M309.04,262.37c-27.4,2.83 -55.21,4.29 -83.36,4.29c-26.07,0 -51.85,-1.27 -77.28,-3.71c3.45,111.74 38.09,199.51 80.31,199.51C271,462.46 305.68,374.38 309.04,262.37z"
+      android:fillColor="#9FD7C5"/>
+  <path
+      android:pathData="M228.71,462.46c42.3,0 76.98,-88.08 80.33,-200.1c-8.13,0.84 -16.29,1.54 -24.49,2.14c-34.49,23.91 -62.37,56.74 -80.57,95.64c-2.03,-0.08 -19.32,1.52 -43.78,-0.75C174.42,421.27 199.78,462.46 228.71,462.46z"
+      android:fillColor="#7BAD9A"/>
+  <path
+      android:pathData="M225.68,359.57c-55.72,0 -108.77,11.37 -156.99,31.91c39.63,43.61 96.63,70.99 160.02,70.99c62.46,0 118.74,-26.58 158.29,-69.1C337.63,371.65 283.07,359.57 225.68,359.57z"
+      android:fillColor="#356D6B"/>
+  <path
+      android:pathData="M295.71,365.7c-22.74,-4.01 -46.14,-6.14 -70.03,-6.14c-21.86,0 -43.3,1.77 -64.21,5.14c14.43,58.9 39.15,97.75 67.24,97.75C256.64,462.46 281.24,424.05 295.71,365.7z"
+      android:fillColor="#6B9984"/>
+  <path
+      android:pathData="M472.83,215.68c-0.17,135.46 -109.47,245.04 -244.12,244.75C228.87,324.96 338.17,215.38 472.83,215.68L472.83,215.68z"
+      android:fillColor="#729800"/>
+  <path
+      android:pathData="M496.74,388.52c-53.7,94.46 -173.32,127.22 -267.16,73.17C283.27,367.23 402.89,334.47 496.74,388.52L496.74,388.52z"
+      android:fillColor="#A5CD27"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M327.61,498.43c-35.31,0 -71.07,-8.89 -103.86,-27.61c-0.42,-0.23 -0.82,-0.48 -1.2,-0.76c-0.92,-0.66 -1.72,-1.43 -2.39,-2.29c-0,-0 -0.01,-0.01 -0.01,-0.01c-0,-0 -0,-0 -0,-0c-0.44,-0.57 -0.84,-1.18 -1.18,-1.84c-0.6,-1.16 -1.01,-2.44 -1.19,-3.8c-0.08,-0.62 -0.12,-1.23 -0.1,-1.84c0.15,-139.46 113.65,-252.87 253.15,-252.87c6.45,0 11.68,5.23 11.68,11.68c0,37.22 -7.9,73.09 -23.49,106.61c-2.72,5.85 -9.67,8.39 -15.52,5.67c-5.85,-2.72 -8.39,-9.67 -5.67,-15.52c12.46,-26.79 19.51,-55.26 21.01,-84.77c-107.23,5.54 -195.27,84.93 -214.09,188.16c64.69,-72.59 173.69,-92.73 261.38,-42.56c2.69,1.54 4.66,4.08 5.47,7.07c0.81,2.99 0.41,6.18 -1.13,8.87C471.68,460.41 400.61,498.43 327.61,498.43zM263.06,463.5c80.43,29.83 172.66,0.7 220.79,-72.18c-78.42,-38.44 -172.7,-16.82 -226.93,49.64l74.4,-20.24c6.23,-1.7 12.64,1.98 14.34,8.2c1.69,6.22 -1.98,12.64 -8.2,14.34L263.06,463.5zM176.35,465.76c-0.94,0 -1.89,-0.11 -2.84,-0.35c-48.72,-12.19 -92.8,-40.71 -124.13,-80.32C17.07,344.25 0,295.09 0,242.91C0,116.45 102.88,13.57 229.34,13.57c0.01,0 0.02,0 0.03,0c97.53,0.01 184.58,61.86 216.62,153.92c2.12,6.09 -1.1,12.75 -7.19,14.87c-6.09,2.12 -12.75,-1.1 -14.87,-7.19c-5.88,-16.89 -13.82,-32.65 -23.49,-47.04c-0.03,0.01 -0.06,0.03 -0.09,0.04c-25.27,11.7 -53.65,20.78 -83.68,26.93c2.89,17.72 4.91,36.42 5.97,55.68c0.35,6.44 -4.58,11.95 -11.02,12.3c-6.41,0.37 -11.95,-4.58 -12.3,-11.02c-1.04,-18.94 -2.97,-36.6 -5.6,-52.87c-23.06,3.49 -46.85,5.3 -70.75,5.3c-19.72,0 -39.21,-1.26 -58.28,-3.74c-3.9,24.69 -6.16,52.32 -6.16,82.18c0,2.43 0.02,4.86 0.05,7.28c25.35,2.19 51.16,3.31 77.1,3.31c2.23,0 4.46,-0.01 6.69,-0.03c0.03,0 0.06,0 0.09,0c6.41,0 11.63,5.17 11.68,11.59c0.05,6.45 -5.14,11.72 -11.59,11.77c-2.28,0.02 -4.57,0.03 -6.86,0.03c-25.66,0 -51.19,-1.07 -76.32,-3.16c1.49,27.56 4.99,53.74 10.3,77.35c8.19,-1.08 16.43,-1.91 24.71,-2.48c6.45,-0.44 12.01,4.41 12.46,10.85c0.44,6.44 -4.41,12.01 -10.85,12.46c-6.83,0.47 -13.63,1.13 -20.39,1.96c2.2,7.37 4.6,14.39 7.2,21.04c2.35,6.01 -0.62,12.78 -6.63,15.13c-6.01,2.35 -12.78,-0.62 -15.13,-6.63c-3.18,-8.14 -6.09,-16.8 -8.71,-25.91c-21.59,4.06 -42.48,9.98 -62.03,17.6c24.57,22.46 54.83,39.15 88.89,47.68c6.26,1.57 10.06,7.91 8.49,14.17C186.34,462.21 181.58,465.76 176.35,465.76zM23.52,251.14c1.87,47.51 19.85,91.46 49.25,125.83c23.02,-9.75 47.94,-17.22 73.79,-22.19c-5.65,-25.56 -9.29,-53.78 -10.71,-83.35C96.78,267.13 58.97,260.32 23.52,251.14zM23.96,227.1c34.96,9.37 72.42,16.33 111.23,20.75c-0.01,-1.64 -0.02,-3.29 -0.02,-4.94c0,-29.89 2.23,-58.93 6.49,-85.8c-30.06,-5.57 -58.79,-14.26 -85.34,-25.86C38.17,159.27 26.64,191.97 23.96,227.1zM168.96,137.74c17.67,2.24 35.73,3.38 54.01,3.38c22.18,0 44.57,-1.73 66.36,-5.01c-13.85,-62.67 -38.55,-99.19 -59.98,-99.19C206.75,36.93 182.41,75.43 168.96,137.74zM70.5,111.91c23.57,9.87 48.92,17.33 75.4,22.24c3.79,-17.82 8.54,-34.4 14.19,-49.29c6.49,-17.11 13.97,-31.36 22.2,-42.51C137.75,52.79 98.73,77.74 70.5,111.91zM276.58,42.41c6.02,8.22 11.99,18.64 17.7,31.78c7.26,16.7 13.33,36.29 18.02,57.83c26.45,-5.44 51.52,-13.21 73.65,-22.98C357.65,76.01 319.36,52.49 276.58,42.41z"/>
+</vector>
