+++ ui/src/main/res/drawable/ic_location_search.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M19,12C19,15.866 15.866,19 12,19M19,12C19,8.134 15.866,5 12,5M19,12H21M12,19C8.134,19 5,15.866 5,12M12,19V21M5,12C5,8.134 8.134,5 12,5M5,12H3M12,5V3M15,12C15,13.657 13.657,15 12,15C10.343,15 9,13.657 9,12C9,10.343 10.343,9 12,9C13.657,9 15,10.343 15,12Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+</vector>
