+++ ui/src/main/res/drawable/ic_order_history.xml
@@ -1,0 +1,33 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="1024"
+    android:viewportHeight="1024">
+  <path
+      android:pathData="M829.3,672.7H190.1c-57,0 -103.2,46.4 -103.2,103.7v60.4h845.5v-60.4c0,-57.3 -46.2,-103.7 -103.2,-103.7z"
+      android:fillColor="#FFCA6C"/>
+  <path
+      android:pathData="M386.3,672.7H190.1c-57,0 -103.2,46.4 -103.2,103.7v60.4h196.2v-60.4c0,-57.3 46.2,-103.7 103.2,-103.7z"
+      android:fillColor="#E5B05C"/>
+  <path
+      android:pathData="M855.4,672.7c0,-191.9 -154.8,-347.4 -345.7,-347.4 -190.9,0 -345.7,155.5 -345.7,347.4"
+      android:fillColor="#FFCA6C"/>
+  <path
+      android:pathData="M567,330.1a346.6,346.6 0,0 0,-57.3 -4.8c-190.9,0 -345.7,155.5 -345.7,347.4h114.5c0,-172.3 124.8,-315.2 288.5,-342.6z"
+      android:fillColor="#E5B05C"/>
+  <path
+      android:pathData="M447.3,211.3a62.7,62.4 90,1 0,124.9 0,62.7 62.4,90 1,0 -124.9,0Z"
+      android:fillColor="#FFCA6C"/>
+  <path
+      android:pathData="M496.3,211.3A62.7,62.7 0,0 1,534.2 153.6,61.9 61.9,0 0,0 509.7,148.6c-34.5,0 -62.4,28.1 -62.4,62.7 0,34.6 28,62.7 62.4,62.7 8.7,0 17,-1.8 24.5,-5a62.8,62.8 0,0 1,-37.9 -57.7z"
+      android:fillColor="#E5B05C"/>
+  <path
+      android:pathData="M783.8,588.6c-5.8,0 -11.2,-3.6 -13.3,-9.4 -32,-89.8 -109.5,-158.1 -202,-178.3 -7.6,-1.7 -12.5,-9.2 -10.8,-16.9s9.2,-12.5 16.8,-10.9a304.9,304.9 0,0 1,137 69.7,308.3 308.3,0 0,1 85.7,126.8 14.2,14.2 0,0 1,-13.3 19zM868.9,745.5a14.1,14.1 0,0 1,-11.3 -5.6c-2.8,-3.7 -9.5,-6.3 -11.7,-6.8a14.2,14.2 0,0 1,6.6 -27.7c1.8,0.4 18.1,4.7 27.6,17.3a14.2,14.2 0,0 1,-11.3 22.8z"
+      android:fillColor="#FFFFFF"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M869.5,665.6c-3.7,-191.4 -156.1,-346.8 -345.6,-354.2v-24.6a76.9,76.9 0,0 0,62.4 -75.6,77.4 77.4,0 0,0 -9.5,-37.2 14.1,14.1 0,0 0,-19.2 -5.6,14.3 14.3,0 0,0 -5.5,19.3c3.9,7.1 6,15.2 6,23.4 0,26.7 -21.7,48.5 -48.3,48.5 -26.6,0 -48.3,-21.8 -48.3,-48.5s21.7,-48.5 48.3,-48.5c6.6,0 13,1.3 19,3.9a14.1,14.1 0,0 0,18.6 -7.5,14.2 14.2,0 0,0 -7.4,-18.7 75.8,75.8 0,0 0,-30.1 -6.2c-42.2,0 -76.6,34.5 -76.6,76.9a77,77 0,0 0,62.4 75.6v24.6a354.8,354.8 0,0 0,-135.2 32.2,14.2 14.2,0 0,0 -7,18.8 14.1,14.1 0,0 0,18.7 7A327.6,327.6 0,0 1,509.7 339.6c178.3,0 324.1,142.1 331.3,319.5a117,117 0,0 0,-11.7 -0.6H190.1c-3.9,0 -7.8,0.2 -11.7,0.6a332.6,332.6 0,0 1,150.2 -265.5,14.2 14.2,0 0,0 4.1,-19.7 14.1,14.1 0,0 0,-19.6 -4.1c-100.3,65.8 -161,176 -163.2,295.9A118,118 0,0 0,72.8 776.4v60.4c0,7.8 6.3,14.2 14.2,14.2h845.5c7.8,0 14.1,-6.4 14.1,-14.2v-60.4c0,-50.8 -32.2,-94.2 -77.1,-110.8zM101.1,822.5v-46.1c0,-49.3 39.9,-89.5 89,-89.5h639.2c44.3,0 81.1,32.6 87.9,75.2H805.9c-7.8,0 -14.1,6.4 -14.1,14.2 0,7.9 6.3,14.2 14.1,14.2h112.5v31.9H101.1z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M756.8,762.2H328.6c-7.8,0 -14.2,6.4 -14.2,14.2 0,7.9 6.3,14.2 14.2,14.2h428.2c7.8,0 14.1,-6.4 14.1,-14.2a14.2,14.2 0,0 0,-14.1 -14.2zM283.1,762.2H162.6c-7.8,0 -14.2,6.4 -14.2,14.2 0,7.9 6.3,14.2 14.2,14.2h120.5c7.8,0 14.2,-6.4 14.2,-14.2 0,-7.8 -6.3,-14.2 -14.1,-14.2z"/>
+</vector>
