+++ ui/src/main/res/drawable/ic_place.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M8.5,13S12,10 12,6.5a3.5,3.5 0,0 0,-7 0C5,10 8.5,13 8.5,13Z"
+      android:strokeWidth="2"
+      android:fillColor="#fa0050"/>
+  <path
+      android:pathData="M12,9h8a1,1 0,0 1,1 1V20a1,1 0,0 1,-1 1H4a1,1 0,0 1,-1 -1V10A1,1 0,0 1,4 9H5"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M13,9l3,8m5,0H3m6,0v4m-0.5,-8S12,10 12,6.5a3.5,3.5 0,0 0,-7 0C5,10 8.5,13 8.5,13Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+</vector>
