+++ ui/src/main/res/drawable/ic_tab_favorite.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M8.962,18.911L9.426,18.322L8.962,18.911ZM12,5.501L11.46,6.021C11.601,6.168 11.796,6.251 12,6.251C12.204,6.251 12.399,6.168 12.54,6.021L12,5.501ZM15.038,18.911L15.503,19.5L15.038,18.911ZM7.001,16.421C6.681,16.158 6.208,16.204 5.945,16.523C5.682,16.843 5.728,17.316 6.047,17.579L7.001,16.421ZM2.342,13.411C2.541,13.775 2.996,13.908 3.36,13.71C3.723,13.511 3.857,13.055 3.658,12.692L2.342,13.411ZM2.75,9.137C2.75,6.986 3.965,5.183 5.624,4.424C7.236,3.687 9.402,3.883 11.46,6.021L12.54,4.981C10.099,2.444 7.264,2.025 5.001,3.06C2.785,4.073 1.25,6.425 1.25,9.137H2.75ZM8.497,19.5C9.01,19.904 9.56,20.334 10.117,20.66C10.674,20.985 11.31,21.25 12,21.25V19.75C11.69,19.75 11.326,19.629 10.874,19.365C10.421,19.101 9.952,18.737 9.426,18.322L8.497,19.5ZM15.503,19.5C16.929,18.375 18.753,17.087 20.183,15.476C21.639,13.836 22.75,11.803 22.75,9.137H21.25C21.25,11.335 20.351,13.028 19.062,14.48C17.747,15.96 16.09,17.127 14.574,18.322L15.503,19.5ZM22.75,9.137C22.75,6.425 21.215,4.073 18.999,3.06C16.736,2.025 13.901,2.444 11.46,4.981L12.54,6.021C14.598,3.883 16.764,3.687 18.376,4.424C20.035,5.183 21.25,6.986 21.25,9.137H22.75ZM14.574,18.322C14.048,18.737 13.579,19.101 13.126,19.365C12.674,19.629 12.31,19.75 12,19.75V21.25C12.69,21.25 13.326,20.985 13.883,20.66C14.441,20.334 14.99,19.904 15.503,19.5L14.574,18.322ZM9.426,18.322C8.63,17.694 7.821,17.096 7.001,16.421L6.047,17.579C6.878,18.262 7.755,18.914 8.497,19.5L9.426,18.322ZM3.658,12.692C3.097,11.666 2.75,10.503 2.75,9.137H1.25C1.25,10.775 1.67,12.183 2.342,13.411L3.658,12.692Z"
+      android:fillColor="#1C274C"/>
+</vector>
