+++ ui/src/main/res/drawable/ic_tab_home.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M9,16C9.85,16.63 10.885,17 12,17C13.115,17 14.15,16.63 15,16"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#1C274C"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M21.636,12.958L21.357,14.895C20.87,18.283 20.626,19.976 19.451,20.988C18.276,22 16.553,22 13.106,22H10.894C7.447,22 5.724,22 4.549,20.988C3.374,19.976 3.13,18.283 2.643,14.895L2.364,12.958C1.985,10.321 1.795,9.002 2.335,7.875C2.876,6.748 4.026,6.062 6.327,4.692L7.712,3.867C9.801,2.622 10.846,2 12,2C13.154,2 14.199,2.622 16.288,3.867L17.673,4.692C19.974,6.062 21.124,6.748 21.665,7.875"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#1C274C"
+      android:strokeLineCap="round"/>
+</vector>
